refactor: define FileEntry type
Start to extract type definitions dynamically
imported from the CLI to explicit definitions
in the core.
