+++ core/types.ts
@@ -1,0 +1,4 @@
+export interface FileEntry {
+  name: string;
+  path: string;
+}
